requirements.txt
fastapi==0.116.1
psycopg2==2.9.10
python-dotenv==1.1.1
sqlmodel==0.0.24
uvicorn==0.35.0